move generateErrorMap
