+++ app/pages/sections/commencement-date/contextCreator.js
@@ -1,6 +1,6 @@
 import manifest from './manifest.json';
 import { baseUrl } from '../../../config';
-import { generateErrorMap } from '../../../helpers/generateErrorMap';
+import { generateErrorMap } from '../../../helpers/contextCreators/generateErrorMap';
 import { generateQuestions } from '../../../helpers/generateQuestions';
 import { generateErrorSummary } from '../../../helpers/generateErrorSummary';
